Unify helper functions for CLI and API
